+++ src/main/java/Capstone/Aeroponics/services/PlantService.java
@@ -8,9 +8,9 @@ import org.hibernate.service.spi.ServiceException;
 import org.springframework.stereotype.Service;
 
 import Capstone.Aeroponics.exception.ResourceNotFoundException;
-import Capstone.Aeroponics.models.RO.PlantRO;
 import Capstone.Aeroponics.models.entities.Plant;
 import Capstone.Aeroponics.models.entities.User;
+import Capstone.Aeroponics.models.request.PlantRO;
 import Capstone.Aeroponics.repositories.PlantRepository;
 import Capstone.Aeroponics.utils.MessageUtils;
 import lombok.RequiredArgsConstructor;
@@ -39,7 +39,7 @@ public class PlantService {
         }
     }
     
-    public Optional<Plant> getById(Long id) {
+    public Optional<Plant> getById(int id) {
         if (Objects.isNull(id)) {
             return Optional.empty();
         }
@@ -47,7 +47,7 @@ public class PlantService {
         return plantRepository.findById(id);
     }
 
-    public Plant getPlantById(Long id) {
+    public Plant getPlantById(int id) {
         try {
             Optional<Plant> plant = getById(id);
 
@@ -73,7 +73,7 @@ public class PlantService {
         }
     }
 
-    public void update(Long id, PlantRO plantRO) {
+    public void update(int id, PlantRO plantRO) {
         try {
             Plant plant = getPlantById(id);
 
@@ -89,7 +89,7 @@ public class PlantService {
         }
     }
 
-    public void delete(Long id) {
+    public void delete(int id) {
         try {
             Plant plant = getPlantById(id);
